h1

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: dhawkgir <dhawkgir@student.21-school.ru    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/05 15:49:20 by dhawkgir          #+#    #+#             */
-/*   Updated: 2021/10/14 16:23:06 by dhawkgir         ###   ########.fr       */
+/*   Updated: 2021/10/17 13:21:18 by dhawkgir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (*(s + i))
